@@ -1,32 +1,39 @@
 {
   "name": "teslaterm",
   "version": "0.1.0",
+  "arch": "win33",
   "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "lint": "tslint -c tslint.json -p tsconfig.json",
+    "build": "cross-var tsc",
+    "watch": "cross-var tsc -w",
+    "lint": "cross-var tslint -c tslint.json -p tsconfig.json",
     "start": "npm run build && npm run start_nobuild",
-    "start_nobuild": "electron dist/compiled/Launch.js --no-sandbox"
+    "start_nobuild": "cross-var electron . --no-sandbox",
+    "install-win": "cd node_modules/@serialport/bindings; prebuild-install --arch x64 --platform win32 --tag-prefix @serialport/bindings@ -r electron -t 7.2.1",
+    "install-linux": "cd node_modules/@serialport/bindings; prebuild-install --arch x64 --platform linux --tag-prefix @serialport/bindings@ -r electron -t 7.2.1",
+    "package-win": "cross-var electron-packager . --overwrite --platform=win32 --arch=x64 --out=dist --prune=true ",
+    "package-linux": "cross-var electron-packager . --overwrite --platform=linux --arch=x64 --out=dist --prune=true "
   },
-  "main": "dist/compiled/Launch.js",
+  "main": "main.js",
   "dependencies": {
     "jquery": "^3.4.1",
     "justgage": "^1.3.5",
     "jzz": "^0.6.9",
     "midi-player-js": "^2.0.5",
-    "serialport": "^8.0.7",
-    "electron": "7.1.11",
-    "typescript": "^3.7.5"
+    "serialport": "^8.0.7"
   },
   "devDependencies": {
-    "@types/serialport": "^8.0.0",
     "@types/chrome": "0.0.78",
-    "@types/jquery": "^3.3.31",
-    "@types/node": "^10.17.14",
+    "@types/jquery": "^3.3.34",
+    "@types/node": "^10.17.18",
+    "@types/serialport": "^8.0.0",
     "@types/w2ui": "^1.4.32",
     "@types/webmidi": "^2.0.4",
-    "electron-rebuild": "^1.10.0",
-    "npm": "6.5.0",
-    "tslint": "^5.10.0"
+    "cross-var": "^1.1.0",
+    "electron": "^7.2.1",
+    "electron-packager": "^14.2.1",
+    "npm": "^6.14.4",
+    "prebuild-install": "^5.3.3",
+    "tslint": "^5.10.0",
+    "typescript": "^3.8.3"
   }
 }
